Move components to page folders if only used once
