+++ src/App.css
@@ -6,6 +6,13 @@
   --nav: #004e92;
 }
 
+#root {
+  margin: 0 auto;
+  line-height: 1.2;
+  height: 100%;
+  width: 100%;
+}
+
 .light-mode {
   --cardBackground: #bbdefb;
   --cardBorder: #78a0cf;
@@ -19,22 +26,45 @@
   --text: #f0edee;
 }
 
-#root {
-  margin: 0 auto;
-  line-height: 1.2;
-  height: 100%;
-  width: 100%;
-}
-
 body {
   background-color: var(--background);
   font-family: "Poppins", "Quicksand", sans-serif;
+}
+
+a {
+  all: unset;
 }
 
 .no-margin {
   margin: 0;
 }
 
-a {
-  all: unset;
+.page-container {
+  width: 100%;
+  max-width: 100%;
+  box-sizing: border-box;
+  padding: 0.4rem;
+  text-align: center;
+}
+
+.main-container {
+  border-radius: 12px;
+  background-color: var(--cardBackground);
+  border: solid 0.1vw var(--cardBorder);
+  color: var(--text);
+  padding: 0.4rem;
+  margin: 0.4rem;
+}
+
+.frosted-item {
+  border-radius: 12px;
+  background: var(--frostedItems);
+  box-shadow: 0 4px 4px rgba(30, 30, 130, 0.3);
+}
+
+.frosted-item p {
+  margin: 0.1rem;
+  white-space: nowrap;
+  text-align: center;
+  width: 100%;
 }
